// packages/ui/ui-kit/src/m3/types/base/base-content.ts
export interface BaseContent<
  TransformerTarget extends HTMLElement,
  _Attributes extends Attributes = Record<string, unknown>,
  Child = string | AllComponentsContent<unknown>,
> {
  children?: Array<Child>;
  attributes?: Merge<Attributes, _Attributes>;
  transformers?: SingleOrArray<Transformer<TransformerTarget>>;
  events?: Events;
}

export type Events<
  Event extends keyof HTMLElementEventMap = keyof HTMLElementEventMap,
> = Partial<Record<Event, (event: HTMLElementEventMap[Event]) => void>>;

export type Attributes = {
  hidden?: boolean;
  slot?: string;
  classes?: Array<string>;
  styles?: Partial<Record<LiteralUnion<CSSProperty, CSSVariable>, string>>;
} & Record<string, unknown>;

export type Transformer<T extends HTMLElement = HTMLElement> = (target: T) => T;

export type CSSVariable = `--${string}`;